requirements.txt
mcstatus
requests
Pillow
deepl
discord-py-interactions
imgur-python